% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bws_cdf}
\alias{bws_cdf}
\title{CDF of the Baumgartner-Weiss-Schindler test under the null.}
\usage{
bws_cdf(b, maxj = 5L, lower_tail = TRUE)
}
\arguments{
\item{b}{a vector of BWS test statistics.}

\item{maxj}{the maximum value of j to take in the approximate computation
of the CDF via equation (2.5). Baumgartner \emph{et. al.} claim that a
value of 3 is sufficient.}

\item{lower_tail}{boolean, when \code{TRUE} returns \eqn{\Psi}{Psi}, otherwise
compute the upper tail, which is more useful for hypothesis tests.}
}
\value{
a vector of the CDF of \eqn{b}, \eqn{\Psi(b)}{Psi(b)}.
}
\description{
Computes the Baumgartner-Weiss-Schindler test statistic under the
null hypothesis of equal distributions.
}
\details{
Given value \eqn{b}, computes the CDF of the BWS statistic under
the null, denoted as \eqn{\Psi(b)}{Psi(b)} by 
Baumgartner \emph{et al.}  The CDF is computed from 
equation (2.5) via numerical quadrature.
}
\examples{

# do it 500 times
set.seed(123)
bvals <- replicate(500, bws_stat(rnorm(50),rnorm(50)))
pvals <- bws_cdf(bvals)
# these should be uniform!
\dontrun{ 
  plot(ecdf(pvals)) 
}

# compare to Table 1 of Baumgartner et al.
bvals <- c(1.933,2.493,3.076,3.880,4.500,5.990)
tab1v <- c(0.9,0.95,0.975,0.990,0.995,0.999)
pvals <- bws_cdf(bvals,lower_tail=TRUE)
show(data.frame(B=bvals,BWS_psi=tab1v,our_psi=pvals))

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
W. Baumgartner, P. Weiss, H. Schindler, 'A Nonparametric Test for the General Two-Sample Problem', 
Biometrics, Vol. 54, No. 3 (Sep., 1998), pp. 1129-1135. 
\url{http://doai.io/10.2307/2533862}
}
\seealso{
bws_stat
}

