% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bws_stat}
\alias{bws_cdf}
\alias{bws_stat}
\title{Perform the Baumgartner-Weiss-Schindler test.}
\usage{
bws_stat(x, y)

bws_cdf(b, maxj = 5L)
}
\arguments{
\item{x}{a vector.}

\item{y}{a vector.}

\item{b}{a vector of BWS test statistics.}

\item{maxj}{the maximum value of j to take in the approximate computation
of the CDF via equation (2.5). Baumgartner \emph{et. al.} claim that a
value of 3 is sufficient.}
}
\value{
return the BWS test statistic, \eqn{b}, or the CDF of \eqn{b} under
the null.
}
\description{
Compute the Baumgartner-Weiss-Schindler test statistic, the CDF under the
null, or perform the hypothesis test.
}
\examples{

 set.seed(1234)
 x <- rnorm(1000)
 y <- rnorm(100)
 bval <- bws_stat(x,y)

 # do it 500 times
 set.seed(123)
 bvals <- replicate(500, bws_stat(rnorm(50),rnorm(50)))
 pvals <- bws_cdf(bvals)
 # these should be uniform!
 \dontrun{ 
   plot(ecdf(pvals)) 
 }

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
W. Baumgartner, P. Weiss, H. Schindler, 'A Nonparametric Test for the General Two-Sample Problem', 
Biometrics, Vol. 54, No. 3 (Sep., 1998), pp. 1129-1135 
\url{http://doai.io/10.2307/2533862}

M. Neuhauser, 'Exact Tests Based on the Baumgartner-Weiss-Schindler Statistic--A Survey', 
Statistical Papers, Vol 46 (2005), pp. 1-30.
\url{http://doai.io/10.1007/BF02762032}

M. Neuhauser, 'One-Sided Two-Sample and Trend Tests Based on a Modified 
Baumgartner-Weiss-Schindler Statistic', J. Nonparametric Statistics, Vol 13
(2001) pp 729-739.
\url{http://doai.io/10.1080/10485250108832874}

H. Murakami, 'K-Sample Rank Test Based on Modified Baumgartner Statistic and
its Power Comparison', J. Jpn. Comp. Statist. Vol 19 (2006), pp. 1-13.
\url{http://doai.io/10.1080/00949655.2010.551516}
}

