% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{murakami_stat_parts}
\alias{murakami_stat}
\alias{murakami_stat_parts}
\title{Compute Murakami's test statistic.}
\usage{
murakami_stat_parts(Parts, flavor = 0L)

murakami_stat(x, y, flavor = 0L)
}
\arguments{
\item{Parts}{a matrix, as output by \code{\link{setparts}}, consisting of
1s and 2s. Each column is a separate test, the rows correspond to the ordered elements
in the grouped set, with no possibility of ties. The 1s and 2s denote which of the two
samples the observation belongs to.}

\item{flavor}{which \sQuote{flavor} of test statistic.}

\item{x}{a vector.}

\item{y}{a vector.}
}
\value{
The BWS test statistic, \eqn{B_j}. For \code{murakami_stat_parts}, a vector of
the test statistics.
}
\description{
Compute one of the modified Baumgartner-Weiss-Schindler test statistics proposed
by Murakami, or Neuhauser.
}
\details{
Given vectors \eqn{X} and \eqn{Y}, computes \eqn{B_{jX}} and \eqn{B_{jY}} 
for some \eqn{j} as described by Murakami and by Neuhauser, returning either their 
their average or their average distance.
The test statistics approximate the weighted square norm of the
difference in CDFs of the two distributions. 

The test statistic is based only on the ranks of the input. If the same
monotonic transform is applied to both vectors, the result should be unchanged.

The various \sQuote{flavor}s of test statistic are:
\describe{
\item{0}{The statistic of Baumgartner-Weiss-Schindler.}
\item{1}{Murakami's \eqn{B_1} statistic, from his 2006 paper.}
\item{2}{Neuhauser's difference statistic, denoted by Murakami as \eqn{B_2} in his 
2012 paper.}
\item{3}{Murakami's \eqn{B_3} statistic, from his 2012 paper.}
\item{4}{Murakami's \eqn{B_4} statistic, from his 2012 paper.}
}
}
\note{
\code{NA} and \code{NaN} not yet dealt with.

this is NYI!
}
\examples{

set.seed(1234)
x <- runif(1000)
y <- runif(100)
bval <- murakami_stat(x,y,1)

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
W. Baumgartner, P. Weiss, H. Schindler, 'A nonparametric test for the general two-sample problem', 
Biometrics 54, no. 3 (Sep., 1998): pp. 1129-1135. 
\url{http://doai.io/10.2307/2533862}

M. Neuhauser, 'Exact tests based on the Baumgartner-Weiss-Schindler Statistic--a survey', 
Statistical Papers 46, no. 1 (2005): pp. 1-30.
\url{http://doai.io/10.1007/BF02762032}

M. Neuhauser, 'One-sided two-sample and trend tests based on a modified 
Baumgartner-Weiss-Schindler statistic', J. Nonparametric Statistics 13, no.
5 (2001): pp 729-739.
\url{http://doai.io/10.1080/10485250108832874}

H. Murakami, 'K-sample rank test based on modified Baumgartner statistic and
its power comparison', J. Jpn. Comp. Statist. 19, no. 1 (2006): pp. 1-13.
\url{http://doai.io/10.5183/jjscs1988.19.1}

H. Murakami, 'Modified Baumgartner Statistics for the two-sample and
multisample problems: a numerical comparison', J. Stat. Comp. and Sim.
82, no. 5 (2012): pp. 711-728.
\url{http://doai.io/10.1080/00949655.2010.551516}

H. Murakami, 'Lepage type statistic based on the modified Baumgartner statistic',
Comp. Stat. & Data Analysis 51 (2007): pp 5061-5067. 
\url{http://doai.io/10.1016/j.csda.2006.04.026}
}
\seealso{
\code{\link{bws_stat}}.
}

