% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWStest.r
\docType{package}
\name{BWStest-package}
\alias{BWStest-package}
\title{Baumgartner Weiss Schindler test of equal distributions.}
\description{
Baumgartner Weiss Schindler test.
}
\section{Background}{


The Baumgartner Weiss Schindler test is a two sample test of the null 
that the samples come from the same probability distribution, similar
to the Kolmogorv-Smirnov, Wilcoxon, and Cramer-Von Mises tests. It is
similar to the Cramer-Von Mises test in that it estimates the
square norm of the difference in CDFs of the two samples. However, the
Baumgartner Weiss Schindler test weights the integral by the variance
of the difference in CDFs, "[emphasizing] the tails of the distributions,
which increases the power of the test for a lot of applications."
}

\section{Legal Mumbo Jumbo}{


BWStest is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\references{
W. Baumgartner, P. Weiss, H. Schindler, 'A nonparametric test for the general two-sample problem', 
Biometrics 54, no. 3 (Sep., 1998): pp. 1129-1135. 
\url{http://doai.io/10.2307/2533862}

M. Neuhauser, 'Exact tests based on the Baumgartner-Weiss-Schindler Statistic--a survey', 
Statistical Papers 46, no. 1 (2005): pp. 1-30.
\url{http://doai.io/10.1007/BF02762032}

M. Neuhauser, 'One-sided two-sample and trend tests based on a modified 
Baumgartner-Weiss-Schindler statistic', J. Nonparametric Statistics 13, no.
5 (2001): pp 729-739.
\url{http://doai.io/10.1080/10485250108832874}

H. Murakami, 'K-sample rank test based on modified Baumgartner statistic and
its power comparison', J. Jpn. Comp. Statist. 19, no. 1 (2006): pp. 1-13.
\url{http://doai.io/10.5183/jjscs1988.19.1}

H. Murakami, 'Modified Baumgartner Statistics for the two-sample and
multisample problems: a numerical comparison', J. Stat. Comp. and Sim.
82, no. 5 (2012): pp. 711-728.
\url{http://doai.io/10.1080/00949655.2010.551516}

H. Murakami, 'Lepage type statistic based on the modified Baumgartner statistic',
Comp. Stat. & Data Analysis 51 (2007): pp 5061-5067. 
\url{http://doai.io/10.1016/j.csda.2006.04.026}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
