% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWStest.r
\docType{package}
\name{BWStest-package}
\alias{BWStest-package}
\title{Baumgartner Weiss Schindler test of equal distributions.}
\description{
Baumgartner Weiss Schindler test.
}
\section{Background}{


more here...
}

\section{Legal Mumbo Jumbo}{


BWStest is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\references{
W. Baumgartner, P. Weiss, H. Schindler, 'A Nonparametric Test for the General Two-Sample Problem', 
Biometrics, Vol. 54, No. 3 (Sep., 1998), pp. 1129-1135. 
\url{http://doai.io/10.2307/2533862}
}
\keyword{package}

