% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWStest.r
\name{BWStest-NEWS}
\alias{BWStest-NEWS}
\title{News for package 'BWStest':}
\description{
News for package 'BWStest'

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cranBWStest}{\CRANpkg{BWStest}}
\newcommand{\BWStest}{\href{https://github.com/shabbychef/BWStest}}
}
\section{\BWStest{} Version 0.2.2 (2018-10-17) }{

\itemize{
\item Package maintenance--no new features.
}
}

\section{\BWStest{} Version 0.2.1 (2017-03-20) }{

\itemize{
\item Package maintenance--no new features.
\item move github figures to location CRAN understands.
\item package initialization mumbo jumbo, see Rcpp issue 636.
}
}

\section{\BWStest{} Version 0.2.0 (2016-04-29) }{

\itemize{
\item Adding Murakami statistics.
}
}

\section{\BWStest{} Version 0.1.0 (2016-04-07) }{

\itemize{
\item First CRAN release.
}
}

\section{\BWStest{} Initial Version 0.0.0 (2016-04-06) }{

\itemize{
\item Start work
}
}

